Update patch set 1
Patch Set 1:

A/B. I'll look into this.

C. I'm with Eddy on this, I really like the transparent look. The three dots should be enough of a hint that it is a dragabble panel and if not the user will realize when they click them.

D. The easiest solution would be to remove the translucent navigation bar in overview mode and let the overview panel fit system views. If we are to keep the translucent navigation bar we will have to programatically add padding below the 3 buttons depending on whether the device has a navigation bar, and even when we do it will seem like excess padding once the panel is expanded.

Patch-set: 1