Update patch set 1
Patch Set 1:

"I would assume that it should be possible if I can drag it up, I should be able to drag it down too." Is this seen anywhere else in system? If you want to go back to homescreen you just tap the screen above the panel.

Patch-set: 1
Label: Code-Review=0