Update patch set 8
Patch Set 8:

I pursued the extending Launcher solution today but it was too messy to dynamically decide which Home Activity to use. I had to add the HOME intent filter to multiple activities, which resulted in a poor user experience.

Instead, I have pulled the CM specific functionality into a new class, GELIntegrationHelper, which is only loaded if the device is running CM. I have tested this build on both CM and AOSP and it is working properly.

Patch-set: 8