Update patch set 2
Patch Set 2:

It's interesting you say that. The previous implementation of the buttons in overview mode used strict padding. The current implementation uses layout_weights to set the correct amount of white space padding between the buttons.

So it's possible to play with this to get a better fit, but we went with this method so that hard coded padding doesn't look odd in tablets and devices with low res.

Patch-set: 2