Update patch set 1
Patch Set 1: Code-Review+1

This makes the workspace centered between the left margin and the hotseats for small tablets (vertical hotseat) in landscape. There was an error in the logic when the hidden searchbar code was written and too much space was allocated for an absent searchbar. 
This works fine on flo and encore and seems to have no effect (as expected) on ovation and tenderloin (without vertical hotseats).

Patch-set: 1
Label: Code-Review=+1