Update patch set 1
Patch Set 1:

Thanks for catching this. I have submitted a new patch that increases the space for the Remove and App Info icons.
http://review.cyanogenmod.org/#/c/65737/
The space on the left side of small (vertical hotseat) is still more symmetric than it is with this revert. 
A problem with ghost icons on the right side in landscape remains but that has been present since the recent Trebuchet 4.4.3 update.

Patch-set: 1
Reviewer: David Marble <2644@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=0