Update patch set 1
Patch Set 1: Code-Review+1

It was pointed out by Eddy Witkamp that the Remove and App Info icons were distorted and cut off with my previous patch.
This increases the space so there is no distortion on small or large tablets. 
There are still two problems with Trebuchet since the 4.4.3 update that this fails to address. 
http://imgur.com/oxmi68A 
illustrates that the Remove and App Info icons are not present in landscape on small tablets if the searchbar is hidden.This image is with my previous patch reverted.
The other problem is ghost icons. Some icons from the screen to the right can appear along the right side in landscape mode.
http://imgur.com/goTjGCK
This happens when you rotate from portrait to landscape.
I have also seen this on a large tablet (tenderloin).

Patch-set: 1
Label: Code-Review=+1