Update patch set 3
Patch Set 3:

So I guess this would be the trade off for having this panel look the same way as it did in the old overview panel. This way, the layout will support longer strings, but will be off-center especially when associated with the handle above it.

The way it currently looks (on mainline) everything will always be centered, but you'll have either line wrap or ellipses.

Unfortunately, this has to be one way or the other.

Patch-set: 3