Update patch set 4
Patch Set 4:

Eddy - added some more spacing in between the buttons so that they take advantage of the some of that extraneous space to the far left and far right. My earlier comment was meant to say that I can't center everything with strict padding because it's not relative as it is with layout_weights. So it kind of looks off center.

Taichi - Tried a fix with 600dp, what'd you think?

Patch-set: 4