Update patch set 4
Patch Set 4:

I think the Google Launcher and this launcher are now centering in the same way.

The container for these buttons is wrapped around the text and centered horizontally. 

When I said they aren't centered, I mean that the space allocated each of the buttons isn't exact. E.g. in the image you posted earlier (with the layout_weight implement of spacing) the white space between of button and on the edges is exactly the same and that each button allocated exactly the same horizontal pixel space, this would continue to scale "exactly" up all screen density buckets.
https://www.dropbox.com/s/p5z824fo7f0h4ze/Screenshot_2014-05-27-19-40-15.jpg

Does this make sense?

Looking into the linewrap on Settings Panel, another place I'm using layout_weights :\

Patch-set: 4