trebuchet: dont remove disabled/not present favorites shortcuts/widgets
This allow to apps disabled or moved to the sdcard don't being its shortcuts and widgets removed from
the workspace, (just grayed is icon) because the external storage is not ready or because the app was
disabled by mistake. The allow the user to decice when to remove any disabled/not present shortcut/widget.

JIRA: CYAN-4296
Issue: https://jira.cyanogenmod.org/browse/CYAN-4296

Change-Id: Ib19e83b5277bb35e16fb6b1d103478a87908f845
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>