Update patch set 5
Patch Set 5: Code-Review+1

This does eliminate the 'ghost icons' on the sides of the landscape screens. The spacing is a bit strange if the search bar is disabled. http://imgur.com/c335ZFe Widgets on the right of the screen overlap the hot seat icons.The icons are centered rather than padded differently on the left and right.I will look into why the padding is ignored but I may not find it. 
There are also no Remove (X) or App info (i) icons on the side if the search bar is disabled. They do appear on the top in portrait. This has been true for a while. 
I never noticed a problem in the AppTray. I keep the labels off.

Patch-set: 5