Update patch set 5
Patch Set 5:

Thanks Rajest. I copied & pasted from other bug. This is the correct one.


Abhisek, I'm sure we don't want to wait for the sdcard to be mounted, because that will delayed the system startup while the system tries to mount the sdcard. In the worst case the sdcard shouldn't be present or be damaged.

Patch-set: 5