Update patch set 5
Patch Set 5:

I'm not so sure adding all this code makes sense either. This is a lot of churn for an edge case. A check at boot would only effect the apps listed as installed externally and in conditions where package manager and sdcard-mount are already racing. With that, I don't think the delay would be that severe (no worse than dexopt).

Patch-set: 5