Update patch set 4
Patch Set 4:

For right now, we have added the custom home screen, but it has no content yet. So, if you enable search panel, you can now see the custom home screen fade in as you scroll off to the left. However, when you have fully scrolled to the new screen and let go, Google Now will be launched. At this time, you can not navigate to the home screen, but when it is ready, that will be an option!

Patch-set: 4
Reviewer: DO NOT USE <9955@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0