Update patch set 2
Patch Set 2:

It works as intended as far as I can see, but the preview images are confusing.

For example, the Comfortable preview led me to believe I would get a 3x3 grid, when it is actually the default 4x4 on my device (i9100).



Is there a way to calculate the actual x by x, based on the device and then show that in the picture?

For example, cozy would be 5x5 for me, so display 5x5 in the preview image (either in dots or in numbers)

Patch-set: 2