Update patch set 10
Patch Set 10:

Please try a nightly that includes this change: http://review.cyanogenmod.org/#/c/65931/. We added a custom activity that displays a translucent fade in before the Google Now Activity is launched. This is closer to the feeling of GEL. 

However, please note that it will not be perfect. The Google Now Activity still takes a moment to launch, and we cannot gradually allow you to swipe directly into the application, as we do not own or control it. This is a technical limitation that means that this will never feel 100% like GEL.

Patch-set: 10
Reviewer: Matt Garnes <9621@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0