Update patch set 1
Patch Set 1: Verified+1

(3 comments)

Good catch, Raj. It seems that opening the app using the globalSearchActivity intent does have that behavior as a side effect. I have added logic to default to using the method we used before (Intent.ACTION_ASSIST) and falling back to globalSearchActivity only if package manager comes up empty for that Intent. 

This way, we'll never crash. If the older version of Google Now is being launched from the fallback Intent, it prompts the user to update right away anyway, so the correct behavior would soon be restored when they do update.

Patch-set: 1