Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

I think this needs some further optimizing. The old behaivior was that the grid of the launcher was only actually changed when the user maid it's choice and went back to the main launcher settings or the even exit the sttings completely. A exeption was the custom grid selection - There the grid was applied when the user made it's choice.

But now we need to always apply the launcher grid directly on user change to draw the grid image dynamically. I think it's better then to drop the "mCustomGridRows" and "mCustomGridColumns" vars completly and improve the general logic a bit.

Some thoughts or even better suggestions about that would be nice :)

Patch-set: 4