Update patch set 4
Patch Set 4:

Glad you're looking to improve this.

Firstly, I don't think you're correct, even if the user set custom values for the grid size, the changes were not applied until the user left the fragment.

Moreover, it's not ideal to apply changes to the grid every time the user makes a changes the selection in this fragment. I believe the old logic was smart enough to apply the the changes only once the user left this fragment.

We can keep this logic and still allow the fragment to redraw the dynamic grid image with the instance variables. "mCustomGridRows" and "mCustomGridColumns" can hold the current num rows and columns so that they can be drawn without the need for applying the changes to the grid directly.

Patch-set: 4