Update patch set 1
Patch Set 1: Code-Review-1

Thanks that you found the root cause. Now I don't want to be a nitpicker, but as you can see from the screenshot, the alpha, although the most visible, is not the only thing that's affected. The scale is too. I suppose a more complete patch would be to also set the scaleX and scaleY to the desired value.
Although I'm a developer alright, I'm completely new to this project, looking just on one source file, so take this with a grain of salt: Wouldn't it be more maintainable (if possible) to get the desired end values from the Animator object itself? It seems that they are passed into its ctor a few lines above the patch, and I dare to suppose they might be available.

Patch-set: 1