Update patch set 1
Patch Set 1:

Good point about the scaling, I didn't realize the scaling wasn't completing either.

Overall this is a bad patch :( The FirstFrameAnimatorHelper is also causing an issue with the FolderIcon behind the opening folder. If you look closely even if I push up the alpha to 1f, you can still see the outline of the FolderIcon behind the opening Folder when this bug happens. When this animation performs correctly, the FolderIcon's alpha is reduced to 0.

Let me rethink this patch, there needs to be a better way to handle this oddness. I'm tempted to just just remove the FirstFrameAnimatorHelper from being added to the animations if the animation scale is below the 1x threshold.

Patch-set: 1