Update patch set 2
Patch Set 2:

Rajesh, I'd love to, but I don't have any android build environment setup at the moment. It might take me some considerable time to do so. Would you, please, build it and share it somewhere for me to grab and test? I suppose it might be easier for the moment.

Patch-set: 2