Update patch set 3
Patch Set 3:

Hi Rajesh, great, the shadow folders are gone. It somehow seemed that I needed to push the animator scale up to 0.4x to recreate the slight animation I was getting before with 0.3x. Otherwise the animation wasn't noticeable, but that's perfect.

Few things though, I see that you added an observer for the scale settings change. I played with it a bit and found out it probably doesn't work. I understand you put it there just to control the result of isAnimatorScaleSafe(), but that alone doesn't seem to work. The animator itself still for some reason won't grab any new settings and the animation is still the same until I kill Trebuchet. Weird. Does it work for you? If it does, then it's probably fine and I missed something in deploying the .apk (perhaps).

Nevertheless, I checked what scale variables matter to the ObjectAnimator (extends ValueAnimator) and have found that it cares just about the Settings.Global.ANIMATOR_DURATION_SCALE, so there's a nice possibility of simplifying the Launcher part to track this single one instead of three.

Patch-set: 3