Update patch set 3
Patch Set 3:

In fact, ValueAnimator has a hidden static function getDurationScale(). That value should be already updated with what's in the Settings (judging from the SDK code.. I don't know why it doesn't work for me), so you might use that in the conditions and not store another copy in the Launcher at all.

Patch-set: 3