Update patch set 3
Patch Set 3:

If you change animation duration in the Development Options you need to restart the process in order for the animation duration change to take affect in that process.
(E.g. if you start with a duration of 1.0x and then change it down to 0.3x you need to restart that application for the animations within that application to respect this new scale). I think the only way to for Trebuchet to respect this to have the observer restart Trebuchet in the event that the animation duration scale changes.

I'm not sure that it makes sense to use a hidden method in the ValueAnimator to get this value. It's a Settings Global field and can easily be accessible inside the launcher app.

I do agree with your comment about only listening for changes in the Animator Duration.

Here's the apk for the new patch: https://www.dropbox.com/s/vc2e48jfz874u91/Trebuchet_70143.apk

Patch-set: 3