Update patch set 5
Patch Set 5:

Fixed the bug you mentioned.

Some interesting things to note:

When you're adjusting the PageScrolls array to accommodate the addition/removal of custom pages you are blinding incrementing the array in the on click.

E.g. if I decide to turn on CMHome and I'm currently set as Off, my mPageScrolls array would read [0, 720, 1440] (in this example I only have one non-custom page).

The reason is because you will add element(1) when I click to GoogleNow and then add element(2) when I click to CMHome.

In any case, the reason we got the NPE you saw was because the of the restore page values we are using when reloading the DynamicGrid. This didn't matter before because we always loaded w/e page we were currently on and that page could never be custom content page. We would be leaving overview mode when the reload kicks off and the custom page isn't visible the overview mode.


Thoughts?

Patch-set: 5