Fix missing workspace screens
Somehow workspace screens are removed on reboot (happens randomly)
However the items in the screens are not. This has 2 effects:
1) The whole screen is missing, which is annoying
2) Readding it by putting the same items there will cause the loader
to complain about items using occupied space (old items vs new ones)
This a) readds the screen if it is missing (but items are still found)
and b) removes items which cannot be placed.

Change-Id: I88d782313eefe1edd982853eff5f2b7b55a5b42a