Update patch set 1
Patch Set 1:

After doing some testing, we found that the OOBE bug which results in half the wallpaper being shown and the other half of the screen is just black is caused by this patch.

Don't have a root cause yet, but this needs to be reverted to resolve this in the short term.

Patch-set: 1