Update patch set 1
Patch Set 1:

(2 comments)

I'd suggest moving all of your screenshot discussion to the comments. If it's not permanent (and img links on the web rarely are), it doesn't really belong in the commit message.  Use a succinct sentence or two to describe the changes. You can move "for example i9300" out of the title and into the message area.

Patch-set: 1