Update patch set 3
Patch Set 3: Verified+1 Code-Review+1

This is a nice addition to make this feature much more visible. Now it really has an effect, whereas I previously could not notice the difference.

I would like to see those numbers get pulled into some kind of array constant or even read from xml possibly to make it cleaner, but that could be an improvement in another patch.

I did notice that when I set the grid size to "condensed," it is not clear what is happening. It kind of looks like the text size changes more than the icon. I think there's not much you can do though when things get that crowded, so I'm ok with the behavior.

Patch-set: 3