Update patch set 4
Patch Set 4:

Well... mainly in n7: http://imgur.com/a/rtn7A

1: n7-small: ok in the homescreen, too tiny in the hotseat
2: n7-big: perfect in the hotseat, a little big (but ok) in the homescreen
3: n7-small: ugly, too tiny
4: n7-big: perfect

5: g3-small: ok in the homescreen, smallish in the hotseat
6: g3-big: ok in the hotseat, a little big in the homescreen
7: g3-small: perfect
8: g3-big: ok


IMO the worst (too much empty space) are 1 and 3, and maybe 5 (only because of the hotseat). The g3 drawer looks great in small mode.

I would revert the icon size of the n7 small profile to the previous values. I would also increase the icon size of the dock icons in both profiles. Why is hotseatIconSize so much lower than iconSize in phones? It may make sense in tablets, since the default homescreen grid is 6x6 and there are 7 icons in the hotseat, but in phones it's 5 in hotseat and 5 (or even more in condensed mode!) in the homescreen.

Patch-set: 4