Update patch set 4
Patch Set 4:

The app drawer button idea would be awesome, but if adding this feature to Trebuchet is out of the question, what about the navbar home button? Correct me if I'm wrong, but I remember having it in the Nexus 5 (activable with a switch in Settings -> Buttons) in the first lollipop versions I compiled. It was way more comfortable than tapping the status bar. Now I have a G3, so I don't know if it's still there for hammerhead.

Patch-set: 4