Update patch set 1
Patch Set 1:

looks like on bacon as the APP button has been moved away to the right abit and little highter, the first pic is perfect except the  distance between APP and SETTING is closer than the others. dunno thats how it looks here and wanted to help and point it out. guess  thats not how it looks on others?

Patch-set: 1