Update patch set 5
Patch Set 5:

Michael: yep, that's what I meant. I'm quite sure double tap and long press for home and recent was configurable in hammerhead, at least in the first CM12 builds. I had enabled double tap on home to sleep, and it worked great. Then I sold the n5, got the g3, was on stock for a few days and when I got to compile CM12 the feature was no longer there. 

Why reserve the navbar button configuration for physical buttons?

Patch-set: 5