Update patch set 2
Patch Set 2: Verified-1 Code-Review+1

This still crashes because we try to use appEntry.componentName.getPackageName() for storing/retrieving icons on lines 239, 297, 302, and 303.

We will have to retrive the ShortcutInfo object with the method you added, but does it make sense to use the icon referenced there? The widget is not technically protected, so should it appear in this list at all?

Patch-set: 2