Update patch set 2
Patch Set 2:

Hm, doesn't make sense that it'd crash there. Those lines are protected by null checks on the componentName. Doesn't crash when I test it.

As for removing it from the list because they are widgets not activities that can be protected, I can go either way on this. Not sure if we should be showing them at all, but in some ways this fragment is just another way to represent the folder so the user may be confused if they can't see this entry in the protected folder list and don't know what happens to it after it gets protected. At the very least it's protecting access to that widget so it's "useful" in a round about way.

Patch-set: 2