Update patch set 1
Patch Set 1:

(2 comments)

Looks pretty good! Awesome job. 

Just as a UX consideration, it looks a little odd to see the scrubber section and the statusbar darken with the apps when fast scrolling. The only way around it that I see, is to make those ui components solid colors. But the transparency affect is also quite nice.

Patch-set: 1