Update patch set 1
Patch Set 1:

(1 comment)

Ok. I see your point, but what happens if you don't bind a widget (but don't remove from the database), and then add shortcuts or widget in the same position that the old widget? IMHO, or we remove (or reorder) the items outside the new grid, or we let the user to remove that items that doesn't fit in its new grid selection. And what happens if the widget fits in the new grid size but is placed in the last col?

Patch-set: 1