Update patch set 1
Patch Set 1:

Looks good. One question, if you were to install or uninstall an application does the sort get re-triggered? If not, this might have been put here as a catch all so it's always sorted when the user enters this view. If it isn't getting sorted, I'd recommend keeping this change and just adding another sort call on addApps or removeApps

Patch-set: 1