Trebuchet: fix Protected apps issue with folders
When there's a folder of shortcuts where all of the items have been set
to protected, the first item in the folder sometimes does not get
removed from the workspace allowing access to the protected application.

When the second-to-last item in a folder is being removed, it was being
replaced with the shortcut automatically, and we were checking against
the folder contents, which was off the workspace at this time, but still
had the shortcut. We need to check whether the last shortcut we are
re-adding back to the workspace is actually protected before adding it.

Change-Id: I9c7dbbd8562ae3e71e198747e6cb1c10417d1f46