Update patch set 2
Patch Set 2:

What about the other icons on the view?  That was my concern.  If I had a calculator widget and other apps.  I guess we would have to force the user to set up their home screen again.  Also, the commit msg was misleading, I was assuming this fixed the scaling issue too :)

Patch-set: 2