Update patch set 5
Patch Set 5:

Quick note about the 9-patches we're using. Decided go with 9patches for the folder backgrounds (for selected and selected states) because we can control that only the edges get scaled not the corners. We do not have this control over vectordrawables (at least not to my knowledge).

Patch-set: 5