Update patch set 2
Patch Set 2:

My test was with bacon without messing with the device display density setting. It ends up being in a different bucket but still uses the same setting. I think the issue is that if it is only defined in values-sw320dp as it was before, that resource will always be used. 

It seems that only when you actually switch to that density bucket does it actually blow up since the original in values/dimens.xml is missing. I'll test on hammerhead too though and report back.

Patch-set: 2