Update patch set 14
Patch Set 14:

First things first, thanks for the bug report there's a fix up for review here: http://review.cyanogenmod.org/#/c/98239/

Thanks for the feedback, I always like to hear what you guys think of what we're trying to do. I honestly don't have all the answers but I can try to portray the inspiration for  these changes and hope to cover some of your concerns.

Now to address the new Folder Icon look. I think what it came down to for us was that the Lollipop redesign for the Folder icon introduced a "vertical stack". This design is fairly ineffective at showing the user what this folder contains. And beyond that, this stack preview requires a lot of fairly interesting and nontrival code to produce. The simpler design the new Folder Icon uses is more basic to be sure, but it also allows the user to see what's in that folder far more easily. Now, we can definitely argue that the user obviously knows what they placed in their folders, despite this the preview should still be effective and showing the folder contents.

As far as expanded mode for folders being full screen, the goal was to add a degree of separation between the folder and the workspace. This degree of separation removes some of the onscreen clutter the user has to deal with when users navigate through the folders. As out of the folder now being less obvious, the new UX stays true to the old exit folder mechanics. Click on an any empty space or the back key will close the folder. And long pressing a short to drag a shortcut out of the folder redraws the folder outline so the user has a clear indication that moving the folder out of that boundary will move them back to the workspace.

Apologies for the long winded response. I'm sure I haven't fully convinced you yet, but I'm open to having more discussions about this. Additionally, @Hayden, you are a far better resource than I for communicating some of the thoughts behind the new UX. So if I misrepresented anything, please correct me.

Patch-set: 14
Reviewer: Rajesh Yengisetty <9528@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0