Update patch set 14
Patch Set 14:

My only complaint is that the folders in the workspace are a bit too iOS-y.

One idea I think would be really cool would be to use the new blur effect (we still need an OSS pixel shader for it) and blur the actual workspace behind the open folder (and dim it as well) instead of only showing the wallpaper. Same thing for the drawer actually. It might be less disorienting.

Patch-set: 14