Update patch set 14
Patch Set 14:

I don't see the proposed toggle as for reverting behavior only, but mainly to accommodate the (obviously differing) tastes. I also highly disagree that we (that is CM) have to be 'different from AOSP', as that sounds like change for the sake of change (and I've outlined my reasoning for believing this actually is a regression instead of an improvement).
I also disagree on using that blur effect everywhere. I realize it's a growing trend in the company, but I don't see it fit the Material Design language at all.

Patch-set: 14