Update patch set 14
Patch Set 14:

The psychology of what's happening here is kind of mind blowing if you step back to think about it. I suspect the Unity guys at Ubuntu spend half the day curled up into the fetal position because of this :)  (I use Gnome 3 for that reason)

By changing the launcher like this, we set out to improve one of the central experiences. By opensourcing it, our hope was to get people excited and interested in something new and fun. 

Since when does an opensource community who runs bleeding edge software want to keep the status quo? What the hell is happening here?

CM has come a long way over the years, from having a zillion settings to become a polished and clean experience. CMParts was cool in it's day but that's not where we have been going. I don't think we've nailed the Launcher UX yet, but the feedback is good and I think we're close.

Patch-set: 14