Update patch set 14
Patch Set 14:

From my perspective, there are many things over the years that I've wanted to do with CM but for whatever reason could not accomplish. For example, there are very few designers in open source so this isn't an area where we could innovate. Now that these kind of resources are available to us, we're being met with resistance and outright aggression for trying to move faster and do things that simply were not possible before. 

There are really three ways to move forward:

1. Develop better process for collaboration with the community and in-house teams. This is ongoing and any input on this is appreciated. This is the ideal path, but I'm worried that the "purist" mentality is too embedded and difficult to break from.

2. Stop opensourcing this kind of work and keep it in our commercial version. This is not ideal because it signifies a complete and utter failure of open-source ideals and the concept that "more eyes are better".

3. Move forward as planned, ignore feedback, piss off contributors. Again, this is obviously bad. I'm also concerned that the handful of people commenting here are not representative of the entirety of the user base, and a basic search around for feedback on forums and social networks show that this is definitely the case- a lot of people love these changes and wonder why the hell we haven't done it before. Maybe add a toggle that says we're unsure of our position on it, which will creep into any new design work done in other parts of the platform.

One thing is clear- for many of us (including myself) CM has never been about "AOSP-for-everything" and that's not where I want the project to be.  People will get pissed off about this, but there are worse things to be hated for than for simply trying something new.

Patch-set: 14