Update patch set 12
Patch Set 12:

(1 comment)

Can we query if the user has a different launcher set as home and not send these stats until Trebuchet is set as the default again? We'd get a daily report of their page and widget count which will never change as long as they don't use Trebuchet. Maybe this is an improvement for another patch though...

Patch-set: 12