Update patch set 1
Patch Set 1:

Think "overdraw" is wrong term here, overdraw is when we have we're asking something to draw on top of another view. Causing us to unnecessarily draw a view that the user doesn't actually see, and cannot interact with. The problem this patch fixes is us causing a scenario where we're causing excessive calls to "dispatchDraw".

Otherwise code looks good

Patch-set: 1