Trebuchet: Fix bug with removing items from folders
The recent folder changes amended the usage of the isDragging
variable to be true only while the drag mechanism is acting on
an open folder. This change breaks behavior when a single item
is being dragged out of the folder. The folder is removed and
replaced with the last item. This shouldn't happen while the
global drag is still in progress, so check drag state using the
drag controller instead of the mDragInProgress instance variable.

Change-Id: I94ea23ca569396ea0241bf0a5af0b23756a13780