Update patch set 2
Patch Set 2: Verified-1

Code looks good and I can confirm that it make it so I can drag the icon out of the folder without it collapsing into one icon now.

However, if you then (without letting the icon go) move it back and drop it onto the folder (without letting the folder expand) the folder disappears and both icons are now gone.

I'm not sure if this patch causes that behavior or if it just repros now that the original bug this patch addressed has been fixed.

Patch-set: 2