Update patch set 4
Patch Set 4:

It seems that folders are already hidden and loading them does not make them visible again. For this reason, I removed the block that sets the alpha to 0f for folders. This would cause them to remain invisible after leaving all apps mode. I haven't pinpointed the reason this happens in code but through testing I've found that the latest patchset works for me.

Patch-set: 4