Update patch set 2
Patch Set 2:

This is an interesting solution, but I wonder if there's a better way to do it.

Either in the OnResume or another activity lifecycle callback the activity is changing the Workspace state, from NORMAL_HIDDEN to NORMAL. Maybe we could trigger closing the folder (with no animation) when that happens as well?

Patch-set: 2