Fade Adjacent Screens: disable this feature for App Drawer (paged)
If enabled on App Drawer it causes some jank when opening the App Drawer
because of the circular reveal anim (the apps flash visible and then go
invisible at the start of the anim).

Also if toggled on and then toggled off, some screens in the App Drawer
become invisible.

Change-Id: Ib7031e9626ce3be56fa0e06ac6394c8e8430ab99