Update patch set 2
Patch Set 2:

PS:

Some code style fixes (if you do "x+y" write "x + y")
Ripped out some dead code
Fixed a bug with the FolderRing which stays in the white accept state and is visible to the user when dragging a shortcut into a folder

TODO:

The font size on the folder edit text needs to drop, on Bacon I couldn't read the full "Tap to name" hint text on a Folder with 2 items.
FolderIcon preview is being cached incorrectly. When you add or remove an item that belongs to the folder icon preview, the next time you open or close the icon you see that the "fake folder icon" preview image is missing the change.

Patch-set: 2