Update patch set 6
Patch Set 6: Code-Review-2

(10 comments)

I am noticing that we are adding a lot of COS code to the CM repo. I was guilty of this as well with my last patch, but I think we should stop before we make the situation worse. We should try to separate everything we can from CM and put it in the COS overlay. Lets set up some time next week to discuss this with Ed.

Patch-set: 6