Add app drawer search bar toggle to settings
This adds the possibility to enable or disable the search bar on top
of the app drawer. In the Launcher class it will be created with the
View.GONE visibility to avoid taking up space if disabled. Per
default it behaves like before and is enabled.

PS3: Make sure the hasSearchBar boolean is set.

Change-Id: I48711aea2b15bf535815d0e453ac33079af26eca