Trebuchet: Never show page indicator in app drawer
When rotation is enabled and we rotate the app launcher, we attempt
to update the indicator alpha state before the view is attached. This
results in the alpha state not applying to the indicator.  Instead,
always set the view to GONE when the app drawer is showing and
the workspace is attached.

Change-Id: Ia2691a12c2cc79ea76a7e91d7d13e94b97df09b5