IconCache: Simplify application of custom titles from STK
Instead of trying to deal with multiple icon-cache conditions,
just apply the custom title to the entry right before using it.

This deals with hot-swap, card exchange, and similar weird issues
that were hitting stale cache entries, and is less aggressive
than outright flushing the cache with every STK refresh.

Change-Id: Iacd7da820f0404b4dc12394d9c9a571bdceadeb0