Comment out debugging activities
These activities cause issues with restricted users even when disabled
by debug_memory_enabled. They appear in the list of activities to be
enabled/disabled when creating a restricted user and if one of them is
not enabled, FallbackHome will be used as default launcher. Fix this
issue commenting out these activities.

BUGBASH-239

Change-Id: Ifc876984c253a48ab5a52d983c0eabaacf732ed5