Disable debugging activities
These activities should be enabled only for debugging the code and
cause issues with restricted users. They appear in the list of
activities to be enabled/disabled for a restricted user and are
disabled by default. If one of them is not enabled, FallbackHome
will be used as default launcher for restricted users. Fix this
issue disabling these activities, which in any case shouldn't be
enabled for regular builds.

BUGBASH-239

Change-Id: Ifc876984c253a48ab5a52d983c0eabaacf732ed5