Don't show market links if market app is protected
Trebuchet is a protected component manager, therefore it can start
protected apps without asking for a PIN or password. When users look
for an app that isn't installed using the search box in the app
drawer, Trebuchet shows a link to start a search using the default
market app. This allows users a way to bypass the protection, so
hide the link if the market is a protected app.

BUGBASH-358

Change-Id: Ic42355cfc1b689818c5d6682a61cf1c98d37bc80