Don't show market links if market app is protected
Trebuchet is a protected component manager, therefore it can start
protected apps without asking for a password. When a user looks for
an app that isn't installed using the search box in the app drawer,
Trebuchet shows a link to start the default market app. This allows
any user to start the market app without inserting the password, even
when protected, so hide the link in this case.

BUGBASH-358

Change-Id: Ic42355cfc1b689818c5d6682a61cf1c98d37bc80