Update patch set 1
Patch Set 1:

The google feed is on the -1 page. If your language is left to right the -1 page is on the left side of your main homescreen. So you need to swipe to the right to see it.

Patch-set: 1
Reviewer: Kim Le <1053@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>